Feature: branch cover test for Point to Polynomial
Before additional tests:
Coverage: 92%
Rows marked as red: (4)
35
39
42
97

These rows are coming after if statements and are never reached as there are no tests that make these if statements true. This is where I added more tests (in the main branch).

To try branch coverage better I partitioned an if statement on row 31 that had two statements into two separate if statements, but both of these were covered by the original tests. 
(Statement 31A was covered by original test 1 and statement 31B was covered by original test 2)

Note: The number of the rows are changed as I add code or comments, so it doesn’t match anymore, but I thought it was helpful for me to keep track over time.